Miscellaneous fixes to make the build process work on other platforms, and to update the schema to account for variances in Oracle configuration.
git-svn-id: https://svn.calendarserver.org/repository/calendarserver/CalendarServer/trunk@7375 e27351fd-9f3e-4f54-a53b-843176b1656c

+++ support/build.sh
@@ -29,6 +29,12 @@ conditional_set () {
   fi;
 }
 
+find_header () {
+  local sysheader="$1"; shift;
+  echo "#include <${sysheader}>" | cc -x c -c - -o /dev/null 2> /dev/null;
+  return "$?";
+}
+
 # Initialize all the global state required to use this library.
 init_build () {
         verbose="";
@@ -138,10 +144,12 @@ apply_patches () {
   fi;
 
   echo "";
-  echo "Removing build directory ${path}/build...";
-  rm -rf "${path}/build";
-  echo "Removing pyc files from ${path}...";
-  find "${path}" -type f -name '*.pyc' -print0 | xargs -0 rm -f;
+  if [ -e "${path}/setup.py" ]; then
+    echo "Removing build directory ${path}/build...";
+    rm -rf "${path}/build";
+    echo "Removing pyc files from ${path}...";
+    find "${path}" -type f -name '*.pyc' -print0 | xargs -0 rm -f;
+  fi;
 }
 
 
@@ -533,6 +541,8 @@ c_dependency () {
   export              PATH="${PATH}:${srcdir}/_root/bin";
   export    C_INCLUDE_PATH="${C_INCLUDE_PATH:-}:${srcdir}/_root/include";
   export   LD_LIBRARY_PATH="${LD_LIBRARY_PATH:-}:${srcdir}/_root/lib";
+  export          CPPFLAGS="${CPPFLAGS:-} -I${srcdir}/_root/include";
+  export           LDFLAGS="${LDFLAGS:-} -L${srcdir}/_root/lib";
   export DYLD_LIBRARY_PATH="${DYLD_LIBRARY_PATH:-}:${srcdir}/_root/lib";
 }
 
@@ -572,6 +582,13 @@ dependencies () {
     :;
   fi;
 
+  if ! find_header ldap.h; then
+    c_dependency -m "ec63f9c2add59f323a0459128846905b" \
+      "OpenLDAP" "openldap-2.4.25" \
+      "http://www.openldap.org/software/download/OpenLDAP/openldap-release/openldap-2.4.25.tgz" \
+      --disable-bdb --disable-hdb;
+  fi;
+
   #
   # Python dependencies
   #
@@ -579,6 +596,9 @@ dependencies () {
   # First, let's make sure that we ourselves are on PYTHONPATH, in case some
   # code (like, let's say, trial) decides to chdir somewhere.
   export PYTHONPATH="${wd}:${PYTHONPATH:-}";
+
+  # Sourceforge mirror hostname.
+  local sf="superb-sea2.dl.sourceforge.net";
 
   local zi="zope.interface-3.3.0";
   py_dependency -m "93668855e37b4691c5c956665c33392c" \
@@ -589,7 +609,7 @@ dependencies () {
   local px="PyXML-${pv}";
   py_dependency -v "${pv}" -m "1f7655050cebbb664db976405fdba209" \
     "PyXML" "xml.dom.ext" "${px}" \
-    "http://superb-sea2.dl.sourceforge.net/project/pyxml/pyxml/${pv}/${px}.tar.gz";
+    "http://${sf}/project/pyxml/pyxml/${pv}/${px}.tar.gz";
 
   local po="pyOpenSSL-0.10";
   py_dependency -v 0.9 -m "34db8056ec53ce80c7f5fc58bee9f093" \
@@ -611,6 +631,13 @@ dependencies () {
   py_dependency -v 0.5 -r 1038 \
     "xattr" "xattr" "xattr" \
     "http://svn.red-bean.com/bob/xattr/releases/xattr-0.6.1/";
+
+  if [ -n "${ORACLE_HOME:-}" ]; then
+      local cx="cx_Oracle-5.1";
+      py_dependency -v 5.1 -m "d2697493a40c9d46c9b7c1c210b61671" \
+          "cx_Oracle" "cx_Oracle" "${cx}" \
+          "http://${sf}/project/cx-oracle/5.1/${cx}.tar.gz";
+  fi;
 
   if [ "${py_version}" != "${py_version##2.5}" ] && ! py_have_module select26; then
     local s26="select26-0.1a3";
